+++ examples/examples_nodejs_test.go
@@ -146,6 +146,16 @@ func TestAccLambdaLayer(t *testing.T) {
 	integration.ProgramTest(t, &test)
 }
 
+func TestAccLambdaContainerImages(t *testing.T) {
+	test := getJSBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			RunUpdateTest: false, // new feature!
+			Dir:           filepath.Join(getCwd(t), "lambda-container-image"),
+		})
+
+	integration.ProgramTest(t, &test)
+}
+
 func TestAccLambdaLayerNewEnums(t *testing.T) {
 	test := getJSBaseOptions(t).
 		With(integration.ProgramTestOptions{